requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
pypdf
streamlit
python-dotenv
qdrant-client
langchain-qdrant